Android allow longer duration when querying without bucket
When executing queryAggregated without a bucket the maximum duration is 2147483647 ms (+/- 25 days) due to the maximum size of an integer.

This change will run the query with duration in seconds if the duration is larger than 2147483647 ms allowing for a maximum of 2147483647 seconds (+/- 68 year). 

It does loose a bit of precision but that's acceptable.